Allow local/separate gateway cluster management

+++ dask-gateway-deploy/6_gateway-shutdown.sh
@@ -1,19 +1,20 @@
+#!/bin/bash
 
 source config.sh
 
-echo ''; kubectl get nodes -o name --namespace ${K8S_NAMESPACE} | xargs -I {} \
-  kubectl drain {} --force --ignore-daemonsets --delete-emptydir-data --grace-period=10
+# Only meaningful for separate Dask gateway deployments.
 
-# gcloud container clusters resize \
-#   ${GKE_CLUSTER_NAME} \
-#   --zone ${GCP_RESOURCE_ZONE} \
-#   --node-pool default-pool \
-#   --num-nodes=0
+if [[ ${ALONGSIDE_JUPYTERHUB} == 0 ]]; then
 
-gcloud container clusters resize \
-  ${GKE_CLUSTER_NAME} \
-  --zone ${GCP_RESOURCE_ZONE} \
-  --node-pool dask-worker-pool \
-  --num-nodes=0
+  echo ''; kubectl get nodes -o name --namespace ${K8S_NAMESPACE} | xargs -I {} \
+    kubectl drain {} --force --ignore-daemonsets --delete-emptydir-data --grace-period=10
 
-echo ''; kubectl get pods --namespace ${K8S_NAMESPACE}
+  gcloud container clusters resize \
+    ${GKE_CLUSTER_NAME} \
+    --zone ${GCP_RESOURCE_ZONE} \
+    --node-pool default-pool \
+    --num-nodes=0
+  
+  echo ''; kubectl get pods --namespace ${K8S_NAMESPACE}
+
+fi
